requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
lunar_python==1.8
aiosmtplib==3.0.1
jinja2==3.1.3
python-multipart==0.0.6
apscheduler==3.10.4
fastapi-mail==1.4.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pillow==10.2.0
